requirements.txt
flask>=3.0.0
gunicorn>=21.2.0
google-cloud-monitoring>=2.20.0
google-api-python-client>=2.140.0
google-auth>=2.35.0
google-auth-httplib2>=0.2.0
httplib2>=0.22.0
